move server files to special folder
