+++ src/server/routes/scenariosRoutes.test.ts
