+++ src/server/routes/scenariosRoutes.ts
